requirements.txt
pandas==0.25.3
numpy==1.17.3
Keras==2.3.1
tensorflow==2.0.0
scikit-learn==0.22.1
pyaml==19.12.0